Put contexts into separate files
