+++ src/App.tsx
@@ -2,12 +2,14 @@ import { Box, Center, ChakraProvider, Flex, Text } from '@chakra-ui/react';
 import { useContext } from 'react';
 import Navbar from './components/Navbar';
 import customizedTheme from './components/theme';
-import { SizeContext } from './contexts/AppContext';
+import CurrentContentContext from './contexts/CurrentContentContext';
+import SizeContext from './contexts/SizeContext';
 import './utilities/standard.css';
 
 /* App proper */
 const App: React.FC = () => {
   const sizes = useContext(SizeContext);
+  const currentContent = useContext(CurrentContentContext);
 
   // TODO: Wrap in `SizeContext.Provider`?
   return (
