+++ src/contexts/CurrentContentContext.ts
@@ -1,0 +1,6 @@
+import { createContext } from 'react';
+import { NavbarButtonIconMap } from '../components/Navbar';
+
+type CurrentContentLiterals = keyof typeof NavbarButtonIconMap;
+
+export default createContext<CurrentContentLiterals>('home');
